deal with unsigned integers correctly
in the name of on disk efficiency (I presume), mysql's goddamn binlog
format apparently abandons whether an integer is signed or not -- even
the mysqlbinlog tool will say stuff like: this integer is -1847483648
or it could be 300,000,000!  This is fine if you're on a release before
5.0, where you can simply coerce mysql into wrapping your negative
numbers around in an unsigned column, but it doesn't work on modern
builds.  I'm not sure what the replication thread is doing about this,
presumably they're not going back through a SQL layer so they don't have
to deal.

In any case, all of this means that I have to hint the signed-ness in
from the captured schema, and then unwrap the unsigned ints in java.

Oh, and why are unsigned integers bad, mr. gosling?  Just abandoning
hope of working with anything outside of java?  Oh well, your beard is
better than mine.